txn: Move pessimistic prewrite action (#8727)
