+++ src/storage/mvcc/mod.rs
@@ -307,7 +307,6 @@ pub fn default_not_found_error(key: Vec<u8>, hint: &str) -> Error {
 pub mod tests {
     use super::*;
     use crate::storage::kv::{Engine, Modify, ScanMode, Snapshot, WriteData};
-    use crate::storage::txn::prewrite;
     use concurrency_manager::ConcurrencyManager;
     use engine_traits::CF_WRITE;
     use kvproto::kvrpcpb::{Context, IsolationLevel};
@@ -362,517 +361,6 @@ pub mod tests {
         let snapshot = engine.snapshot(&ctx).unwrap();
         let mut reader = MvccReader::new(snapshot, None, true, IsolationLevel::Si);
         assert!(reader.get(&Key::from_raw(key), ts.into(), false).is_err());
-    }
-
-    pub fn try_pessimistic_prewrite_check_not_exists<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) -> Result<()> {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let ts = ts.into();
-        let cm = ConcurrencyManager::new(ts);
-        let mut txn = MvccTxn::new(snapshot, ts, true, cm);
-
-        txn.pessimistic_prewrite(
-            Mutation::CheckNotExists(Key::from_raw(key)),
-            pk,
-            &None,
-            false,
-            0,
-            TimeStamp::default(),
-            0,
-            TimeStamp::default(),
-            false,
-        )?;
-        Ok(())
-    }
-
-    pub fn must_prewrite_put_impl<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        secondary_keys: &Option<Vec<Vec<u8>>>,
-        ts: TimeStamp,
-        is_pessimistic_lock: bool,
-        lock_ttl: u64,
-        for_update_ts: TimeStamp,
-        txn_size: u64,
-        min_commit_ts: TimeStamp,
-        pipelined_pessimistic_lock: bool,
-    ) {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let cm = ConcurrencyManager::new(ts);
-        let mut txn = MvccTxn::new(snapshot, ts, true, cm);
-
-        let mutation = Mutation::Put((Key::from_raw(key), value.to_vec()));
-        if for_update_ts.is_zero() {
-            prewrite(
-                &mut txn,
-                mutation,
-                pk,
-                &secondary_keys,
-                false,
-                lock_ttl,
-                txn_size,
-                min_commit_ts,
-            )
-            .unwrap();
-        } else {
-            txn.pessimistic_prewrite(
-                mutation,
-                pk,
-                &secondary_keys,
-                is_pessimistic_lock,
-                lock_ttl,
-                for_update_ts,
-                txn_size,
-                min_commit_ts,
-                pipelined_pessimistic_lock,
-            )
-            .unwrap();
-        }
-        write(engine, &ctx, txn.into_modifies());
-    }
-
-    pub fn must_prewrite_put<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) {
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            &None,
-            ts.into(),
-            false,
-            0,
-            TimeStamp::default(),
-            0,
-            TimeStamp::default(),
-            false,
-        );
-    }
-
-    pub fn must_pessimistic_prewrite_put<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            &None,
-            ts.into(),
-            is_pessimistic_lock,
-            0,
-            for_update_ts.into(),
-            0,
-            TimeStamp::default(),
-            false,
-        );
-    }
-
-    pub fn must_pipelined_pessimistic_prewrite_put<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            &None,
-            ts.into(),
-            is_pessimistic_lock,
-            0,
-            for_update_ts.into(),
-            0,
-            TimeStamp::default(),
-            true,
-        );
-    }
-
-    pub fn must_pessimistic_prewrite_put_with_ttl<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-        lock_ttl: u64,
-    ) {
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            &None,
-            ts.into(),
-            is_pessimistic_lock,
-            lock_ttl,
-            for_update_ts.into(),
-            0,
-            TimeStamp::default(),
-            false,
-        );
-    }
-
-    pub fn must_prewrite_put_for_large_txn<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        ttl: u64,
-        for_update_ts: impl Into<TimeStamp>,
-    ) {
-        let lock_ttl = ttl;
-        let ts = ts.into();
-        let min_commit_ts = (ts.into_inner() + 1).into();
-        let for_update_ts = for_update_ts.into();
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            &None,
-            ts,
-            !for_update_ts.is_zero(),
-            lock_ttl,
-            for_update_ts,
-            0,
-            min_commit_ts,
-            false,
-        );
-    }
-
-    pub fn must_prewrite_put_async_commit<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        secondary_keys: &Option<Vec<Vec<u8>>>,
-        ts: impl Into<TimeStamp>,
-        min_commit_ts: impl Into<TimeStamp>,
-    ) {
-        assert!(secondary_keys.is_some());
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            secondary_keys,
-            ts.into(),
-            false,
-            100,
-            TimeStamp::default(),
-            0,
-            min_commit_ts.into(),
-            false,
-        );
-    }
-
-    pub fn must_pessimistic_prewrite_put_async_commit<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        secondary_keys: &Option<Vec<Vec<u8>>>,
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-        min_commit_ts: impl Into<TimeStamp>,
-    ) {
-        assert!(secondary_keys.is_some());
-        must_prewrite_put_impl(
-            engine,
-            key,
-            value,
-            pk,
-            secondary_keys,
-            ts.into(),
-            is_pessimistic_lock,
-            100,
-            for_update_ts.into(),
-            0,
-            min_commit_ts.into(),
-            false,
-        );
-    }
-
-    fn must_prewrite_put_err_impl<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-        pipelined_pessimistic_lock: bool,
-    ) -> Error {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let for_update_ts = for_update_ts.into();
-        let cm = ConcurrencyManager::new(for_update_ts);
-        let mut txn = MvccTxn::new(snapshot, ts.into(), true, cm);
-        let mutation = Mutation::Put((Key::from_raw(key), value.to_vec()));
-
-        if for_update_ts.is_zero() {
-            prewrite(
-                &mut txn,
-                mutation,
-                pk,
-                &None,
-                false,
-                0,
-                0,
-                TimeStamp::default(),
-            )
-            .unwrap_err()
-        } else {
-            txn.pessimistic_prewrite(
-                mutation,
-                pk,
-                &None,
-                is_pessimistic_lock,
-                0,
-                for_update_ts,
-                0,
-                TimeStamp::default(),
-                pipelined_pessimistic_lock,
-            )
-            .unwrap_err()
-        }
-    }
-
-    pub fn must_prewrite_put_err<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) -> Error {
-        must_prewrite_put_err_impl(engine, key, value, pk, ts, TimeStamp::zero(), false, false)
-    }
-
-    pub fn must_pessimistic_prewrite_put_err<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) -> Error {
-        must_prewrite_put_err_impl(
-            engine,
-            key,
-            value,
-            pk,
-            ts,
-            for_update_ts,
-            is_pessimistic_lock,
-            false,
-        )
-    }
-
-    pub fn must_pipelined_pessimistic_prewrite_put_err<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        value: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) -> Error {
-        must_prewrite_put_err_impl(
-            engine,
-            key,
-            value,
-            pk,
-            ts,
-            for_update_ts,
-            is_pessimistic_lock,
-            true,
-        )
-    }
-
-    fn must_prewrite_delete_impl<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let for_update_ts = for_update_ts.into();
-        let cm = ConcurrencyManager::new(for_update_ts);
-        let mut txn = MvccTxn::new(snapshot, ts.into(), true, cm);
-        let mutation = Mutation::Delete(Key::from_raw(key));
-
-        if for_update_ts.is_zero() {
-            prewrite(
-                &mut txn,
-                mutation,
-                pk,
-                &None,
-                false,
-                0,
-                0,
-                TimeStamp::default(),
-            )
-            .unwrap();
-        } else {
-            txn.pessimistic_prewrite(
-                mutation,
-                pk,
-                &None,
-                is_pessimistic_lock,
-                0,
-                for_update_ts,
-                0,
-                TimeStamp::default(),
-                false,
-            )
-            .unwrap();
-        }
-        engine
-            .write(&ctx, WriteData::from_modifies(txn.into_modifies()))
-            .unwrap();
-    }
-
-    pub fn must_prewrite_delete<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) {
-        must_prewrite_delete_impl(engine, key, pk, ts, TimeStamp::zero(), false);
-    }
-
-    pub fn must_pessimistic_prewrite_delete<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        must_prewrite_delete_impl(engine, key, pk, ts, for_update_ts, is_pessimistic_lock);
-    }
-
-    fn must_prewrite_lock_impl<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let for_update_ts = for_update_ts.into();
-        let cm = ConcurrencyManager::new(for_update_ts);
-        let mut txn = MvccTxn::new(snapshot, ts.into(), true, cm);
-
-        let mutation = Mutation::Lock(Key::from_raw(key));
-        if for_update_ts.is_zero() {
-            prewrite(
-                &mut txn,
-                mutation,
-                pk,
-                &None,
-                false,
-                0,
-                0,
-                TimeStamp::default(),
-            )
-            .unwrap();
-        } else {
-            txn.pessimistic_prewrite(
-                mutation,
-                pk,
-                &None,
-                is_pessimistic_lock,
-                0,
-                for_update_ts,
-                0,
-                TimeStamp::default(),
-                false,
-            )
-            .unwrap();
-        }
-        engine
-            .write(&ctx, WriteData::from_modifies(txn.into_modifies()))
-            .unwrap();
-    }
-
-    pub fn must_prewrite_lock<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) {
-        must_prewrite_lock_impl(engine, key, pk, ts, TimeStamp::zero(), false);
-    }
-
-    pub fn must_prewrite_lock_err<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-    ) {
-        let ctx = Context::default();
-        let snapshot = engine.snapshot(&ctx).unwrap();
-        let ts = ts.into();
-        let cm = ConcurrencyManager::new(ts);
-        let mut txn = MvccTxn::new(snapshot, ts, true, cm);
-
-        assert!(prewrite(
-            &mut txn,
-            Mutation::Lock(Key::from_raw(key)),
-            pk,
-            &None,
-            false,
-            0,
-            0,
-            TimeStamp::default(),
-        )
-        .is_err());
-    }
-
-    pub fn must_pessimistic_prewrite_lock<E: Engine>(
-        engine: &E,
-        key: &[u8],
-        pk: &[u8],
-        ts: impl Into<TimeStamp>,
-        for_update_ts: impl Into<TimeStamp>,
-        is_pessimistic_lock: bool,
-    ) {
-        must_prewrite_lock_impl(engine, key, pk, ts, for_update_ts, is_pessimistic_lock);
     }
 
     pub fn must_acquire_pessimistic_lock_impl<E: Engine>(
