+++ src/storage/mvcc/reader/scanner/backward.rs
@@ -434,7 +434,7 @@ mod tests {
     use super::*;
     use crate::storage::kv::{Engine, TestEngineBuilder};
     use crate::storage::mvcc::tests::*;
-    use crate::storage::txn::tests::must_commit;
+    use crate::storage::txn::tests::{must_commit, must_prewrite_delete, must_prewrite_put};
     use crate::storage::Scanner;
     use kvproto::kvrpcpb::Context;
